refactor/ abstarct out AppName and use AppHeader in pages after logging in
