+++ src/components/CustomerQuota/CustomerQuotaScreen.tsx
@@ -12,7 +12,7 @@ import { DarkButton } from "../Layout/Buttons/DarkButton";
 import { color, size, borderRadius, fontSize } from "../../common/styles";
 import { postTransaction, Quota, Transaction } from "../../services/quota";
 import { useAuthenticationContext } from "../../context/auth";
-import { AppName } from "../Layout/AppName";
+import { AppHeader } from "../Layout/AppHeader";
 import { SecondaryButton } from "../Layout/Buttons/SecondaryButton";
 import { AppText } from "../Layout/AppText";
 import { TopBackground } from "../Layout/TopBackground";
@@ -384,7 +384,7 @@ export const CustomerQuotaScreen: FunctionComponent<NavigationProps> = ({
       <TopBackground mode={config.appMode} />
       <View style={styles.content}>
         <View style={styles.headerText}>
-          <AppName mode={config.appMode} />
+          <AppHeader mode={config.appMode} />
         </View>
 
         {hasPurchased ? (
